fix-attn (#430)

+++ zeromodels/models/gpt_oss/gpt_oss_layers.py
@@ -268,7 +268,7 @@ class GptOssAttention(layers.Layer):
 
         attn = ops.matmul(q, ops.transpose(k, (0, 1, 3, 2))) * self.scaling
         if attention_mask is not None:
-            attn = attn + attention_mask
+            attn = attn + ops.cast(attention_mask, attn.dtype)
         sink_col = ops.zeros_like(attn[..., :1]) + ops.reshape(
             self.sinks, (1, self.num_heads, 1, 1)
         )
