+++ zeromodels/models/qwen3_5/qwen3_5_layers.py
@@ -251,7 +251,7 @@ class Qwen3_5Attention(layers.Layer):
 
         attn = ops.matmul(query, ops.transpose(key, (0, 1, 3, 2))) * self.scaling
         if attention_mask is not None:
-            attn = attn + attention_mask
+            attn = attn + ops.cast(attention_mask, attn.dtype)
         attn = ops.cast(ops.softmax(ops.cast(attn, "float32"), axis=-1), query.dtype)
         out = ops.matmul(attn, value)
         out = ops.reshape(
